refactor: move plus, minus, intersect to abstract class
