+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetImpl.kt
@@ -1,7 +1,5 @@
 package xyz.lbres.kotlinutils.set.multiset
 
-import kotlin.math.min
-
 /**
  * Set implementation that allows multiple occurrences of the same value.
  */
@@ -78,146 +76,8 @@ internal class MultiSetImpl<E> : AbstractMultiSet<E> {
         hashCodes = getCurrentHashCodes()
     }
 
-//    /**
-//     * Determine if an element is contained in the current set.
-//     *
-//     * @param element [E]
-//     * @return [Boolean]: `true` if [element] is in the set, `false` otherwise
-//     */
-//    override fun contains(element: E): Boolean {
-//        updateValues()
-//        return countsMap.contains(element)
-//    }
-//
-//    /**
-//     * Determine if all elements in a collection are contained in the current set.
-//     * If [elements] contains multiple occurrences of the same value, the function will check if this set contains at least as many occurrences as [elements].
-//     *
-//     * @param elements [Collection]<E>
-//     * @return [Boolean]: `true` if the current set contains at least as many occurrences of each value as [elements], `false` otherwise
-//     */
-//    override fun containsAll(elements: Collection<E>): Boolean {
-//        if (elements.isEmpty()) {
-//            return true
-//        }
-//
-//        updateValues()
-//        val newSet = MultiSetImpl(elements)
-//        return newSet.distinctValues.all {
-//            getCountWithoutUpdate(it) > 0 && newSet.getCountWithoutUpdate(it) <= getCountWithoutUpdate(it)
-//        }
-//    }
-
-    /**
-     * Create a new MultiSet with values that are in this set but not the other set.
-     * If there are multiple occurrences of a value, the number of occurrences in the other set will be subtracted from the number in this MultiSet.
-     *
-     * @param other [MultiSet]<E>: MultiSet to subtract from current
-     * @return [MultiSet]<E>: MultiSet containing the items in this MultiSet but not the other
-     */
-    override operator fun minus(other: MultiSet<E>): MultiSet<E> {
-        updateValues()
-
-        val newCounts = distinctValues.associateWith {
-            getCountWithoutUpdate(it) - other.getCountOf(it)
-        }.filter { it.value > 0 }
-
-        return MultiSetImpl(newCounts)
-    }
-
-    /**
-     * Create a new MultiSet with all values from both sets.
-     * If there are multiple occurrences of a value, the number of occurrences in the other set will be added to the number in this MultiSet.
-     *
-     * @param other [MultiSet]<E>: MultiSet to add to current
-     * @return [MultiSet]<E>: MultiSet containing all values from both MultiSets
-     */
-    override operator fun plus(other: MultiSet<E>): MultiSet<E> {
-        updateValues()
-        val allValues = distinctValues + other.distinctValues
-
-        val newCounts = allValues.associateWith {
-            getCountWithoutUpdate(it) + other.getCountOf(it)
-        }
-
-        return MultiSetImpl(newCounts)
-    }
-
-    /**
-     * Create a new MultiSet with values that are shared between the sets.
-     * If there are multiple occurrences of a value, the smaller number of occurrences will be used.
-     *
-     * @param other [MultiSet]<E>: MultiSet to intersect with current
-     * @return [MultiSet]<E>: MultiSet containing only values that are in both MultiSets
-     */
-    override infix fun intersect(other: MultiSet<E>): MultiSet<E> {
-        updateValues()
-        val allValues = distinctValues + other.distinctValues
-
-        val newCounts = allValues.associateWith {
-            val count = getCountWithoutUpdate(it)
-            val otherCount = other.getCountOf(it)
-            min(count, otherCount)
-        }.filter { it.value > 0 }
-
-        return MultiSetImpl(newCounts)
-    }
-
-//    /**
-//     * If the current set contains 0 elements.
-//     *
-//     * @return [Boolean]: `true` if the set contains 0 elements, `false` otherwise
-//     */
-//    override fun isEmpty(): Boolean = countsMap.isEmpty()
-
-//    /**
-//     * Get the number of occurrences of a given element.
-//     *
-//     * @param element [E]
-//     * @return [Int]: the number of occurrences of [element]. 0 if the element does not exist.
-//     */
-//    override fun getCountOf(element: E): Int {
-//        updateValues()
-//        return getCountWithoutUpdate(element)
-//    }
-
-    /**
-     * Get number of occurrences of a given element, without updating the values in the counts map.
-     * Assumes that values have already been updated.
-     *
-     * @param element [E]
-     * @return [Int]: the number of occurrences of [element]. 0 if the element does not exist.
-     */
-    //override fun getCountWithoutUpdate(element: E): Int = countsMap.getOrDefault(element, 0)
-
-    /**
-     * If two MultiSets contain the same elements, with the same number of occurrences per element.
-     *
-     * @param other [Any]?
-     * @return [Boolean]: `true` if [other] is a non-null MultiSet which contains the same values as the current set, `false` otherwise
-     */
-    override fun equals(other: Any?): Boolean {
-        if (other == null || other !is MultiSet<*>) {
-            return false
-        }
-
-        updateValues()
-        return try {
-            @Suppress("UNCHECKED_CAST")
-            other as MultiSet<E>
-
-            // if (other is MultiSetImpl<*>) {
-                // other.up
-                //other.updateValues()
-                //return countsMap == other.countsMap
-            // }
-
-            // less efficient equality check
-            val otherDistinct = other.distinctValues
-            return distinctValues == otherDistinct && distinctValues.all { getCountWithoutUpdate(it) == other.getCountOf(it) }
-        } catch (_: Exception) {
-            false
-        }
+    override fun createFromCountsMap(counts: Map<E, Int>): MultiSet<E> {
+        return MultiSetImpl(counts)
     }
 
     /**
@@ -272,17 +132,5 @@ internal class MultiSetImpl<E> : AbstractMultiSet<E> {
      */
     override fun iterator(): Iterator<E> = initialElements.toList().iterator()
 
-//    /**
-//     * Get a string representation of the set.
-//     *
-//     * @return [String]
-//     */
-//    override fun toString(): String {
-//        updateValues()
-//        return string
-//    }
-
-    override fun hashCode(): Int {
-        return listOf(javaClass.name, initialElements).hashCode()
-    }
+    override fun hashCode(): Int = listOf(javaClass.name, initialElements).hashCode()
 }
